Moved some components from shared to Dialogs folder
