+++ src/components/common/Dialogs/Messages/PermissionRequiredInfo/PermissionRequiredInfo.tsx
@@ -5,8 +5,9 @@ import { ColonyRole } from '@colony/colony-js';
 import Heading from '~shared/Heading';
 import PermissionsLabel from '~shared/PermissionsLabel';
 import Icon from '~shared/Icon';
-import styles from './PermissionRequiredInfo.css';
 import { Tooltip } from '~shared/Popover';
+
+import styles from './PermissionRequiredInfo.css';
 
 const MSG = defineMessages({
   title: {
