+++ src/components/common/Dialogs/Messages/index.ts
@@ -1,0 +1,3 @@
+export { default as NoPermissionMessage } from './NoPermissionMessage';
+export { default as CannotCreateMotionMessage } from './CannotCreateMotionMessage';
+export { default as PermissionRequiredInfo } from './PermissionRequiredInfo';
