+++ src/components/common/Dialogs/TokenManagementDialog/TokenManagementDialogForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TokenManagementDialogForm';
