+++ src/components/common/Dialogs/TokenManagementDialog/TokenManagementDialogForm/tokens/tokenList.xdai.json
