--- src/components/shared/DomainFundSelectorSection/helpers.ts
@@ -1,11 +1,0 @@
-import { Domain } from '~types';
-import { sortBy } from '~utils/lodash';
-
-export const getDomainOptions = (colonyDomains: Domain[]) =>
-  sortBy(
-    colonyDomains.map((domain) => ({
-      value: domain?.nativeId || '',
-      label: domain?.metadata?.name || `Domain #${domain?.nativeId}`,
-    })) || [],
-    ['value'],
-  );
